Finished login screen
The login screen is nothing more than to allow the band members
to load their personal settings.

It uses devise in the backend just so its easy to know who is
logged into that session and retrieve the correct info.

Users will have to be created manually as of right now.
email will be "name@lof.com"
password will be "password"

This can be fine tuned for easier customability later since it's
not hugely important in the context of our personal band.  As
it is scaled out and possibly released, this will need to be
addressed.