Significantly re-did all the javascript
Now it is much easier to add in new components, just create the
component and insert it into GUI.add_component().

slider_vertical is now the template for adding new components.
There are certain functions that components must have in order
for them to be loaded into the GUI properly.

Streamlined the actioncable process.

Removed much unnecessary js files and code.